+++ src/main/java/com/wakeUpTogetUp/togetUp/api/avatar/application/strategy/MakeSpeechStrategy.java
@@ -2,7 +2,7 @@ package com.wakeUpTogetUp.togetUp.api.avatar.application.strategy;
 
 import com.wakeUpTogetUp.togetUp.api.avatar.domain.AvatarSpeech;
 
-public interface SpeechStrategy {
+public interface MakeSpeechStrategy {
 
     String makeSpeech(AvatarSpeech avatarSpeech);
 }
